requirements.txt
python-docx>=1.1.0
tiktoken>=0.5.1
openai>=1.25.0
backoff>=2.2.1
pydantic>=2.0.0
tqdm>=4.66.0
python-dotenv>=1.0.0 